+++ server/src/test/java/org/opensearch/action/admin/cluster/remotestore/stats/RemoteStoreStatsTestHelper.java
@@ -46,6 +46,7 @@ public class RemoteStoreStatsTestHelper {
             0,
             0,
             0,
+            10,
             createZeroDirectoryFileTransferStats()
         );
     }
@@ -53,6 +54,7 @@ public class RemoteStoreStatsTestHelper {
     static RemoteSegmentTransferTracker.Stats createStatsForNewReplica(ShardId shardId) {
         return new RemoteSegmentTransferTracker.Stats(
             shardId,
+            0,
             0,
             0,
             0,
@@ -96,16 +98,17 @@ public class RemoteStoreStatsTestHelper {
             0,
             0,
             100,
+            10,
             createSampleDirectoryFileTransferStats()
         );
     }
 
     static DirectoryFileTransferTracker.Stats createSampleDirectoryFileTransferStats() {
-        return new DirectoryFileTransferTracker.Stats(10, 0, 10, 12345, 5, 5, 5);
+        return new DirectoryFileTransferTracker.Stats(10, 0, 10, 12345, 5, 5, 5, 10);
     }
 
     static DirectoryFileTransferTracker.Stats createZeroDirectoryFileTransferStats() {
-        return new DirectoryFileTransferTracker.Stats(0, 0, 0, 0, 0, 0, 0);
+        return new DirectoryFileTransferTracker.Stats(0, 0, 0, 0, 0, 0, 0, 0);
     }
 
     static ShardRouting createShardRouting(ShardId shardId, boolean isPrimary) {